+++ plugs/index/space_lua.test.ts
@@ -1,36 +1,41 @@
 import { parseMarkdown } from "../../client/markdown_parser/parser.ts";
 import { createMockSystem } from "../../plug-api/system_mock.ts";
 import type { PageMeta } from "@silverbulletmd/silverbullet/type/index";
-import { indexPage } from "./page.ts";
 import { extractFrontMatter } from "./frontmatter.ts";
 import { assertEquals } from "@std/assert";
+import { indexSpaceLua } from "./space_lua.ts";
 
 const testPage = `
----
-hello: attribute
----
-[hello2: 12]
+Hello
+\`\`\`space-lua
+function sup()
+end
+\`\`\`
 
-# Hello world!
-
+\`\`\`space-lua
+-- priority: 10
+function sup()
+end
+\`\`\`
 `.trim();
 
-Deno.test("Test page indexing", async () => {
+Deno.test("Test space lua indexing", async () => {
   createMockSystem();
+
   const tree = parseMarkdown(testPage);
   const frontmatter = extractFrontMatter(tree);
 
   const pageMeta: PageMeta = {
-    ref: "test",
-    name: "test",
+    ref: "folder/test",
+    name: "folder/test",
     tag: "page",
     created: "",
     lastModified: "",
     perm: "rw",
   };
 
-  const [pm] = await indexPage(pageMeta, frontmatter, tree);
-  assertEquals(pm.hello, "attribute");
-  assertEquals(pm.hello2, 12);
-  assertEquals(pm.itags, ["page"]);
+  const objects = await indexSpaceLua(pageMeta, frontmatter, tree);
+  assertEquals(objects.length, 2);
+  assertEquals(objects[0].priority, undefined);
+  assertEquals(objects[1].priority, 10);
 });