+++ plugs/index/table.test.ts
@@ -1,22 +1,20 @@
 import { parseMarkdown } from "../../client/markdown_parser/parser.ts";
 import { createMockSystem } from "../../plug-api/system_mock.ts";
 import type { PageMeta } from "@silverbulletmd/silverbullet/type/index";
-import { indexPage } from "./page.ts";
 import { extractFrontMatter } from "./frontmatter.ts";
 import { assertEquals } from "@std/assert";
+import { indexTables } from "./table.ts";
 
 const testPage = `
----
-hello: attribute
----
-[hello2: 12]
-
-# Hello world!
-
+| name | age |
+|----------|----------|
+| Frank | 20 |
+| Fred | 21 |
 `.trim();
 
-Deno.test("Test page indexing", async () => {
+Deno.test("Test table indexing", async () => {
   createMockSystem();
+
   const tree = parseMarkdown(testPage);
   const frontmatter = extractFrontMatter(tree);
 
@@ -29,8 +27,10 @@ Deno.test("Test page indexing", async () => {
     perm: "rw",
   };
 
-  const [pm] = await indexPage(pageMeta, frontmatter, tree);
-  assertEquals(pm.hello, "attribute");
-  assertEquals(pm.hello2, 12);
-  assertEquals(pm.itags, ["page"]);
+  const datas = await indexTables(pageMeta, frontmatter, tree);
+  assertEquals(datas.length, 2);
+  assertEquals(datas[0].name, "Frank");
+  assertEquals(datas[0].age, "20");
+  assertEquals(datas[1].name, "Fred");
+  assertEquals(datas[1].age, "21");
 });